Pull Request

+++ app/src/main/java/us/shandian/giga/ui/common/ProgressDrawable.java
@@ -9,8 +9,9 @@ import android.graphics.Rect;
 import android.graphics.drawable.Drawable;
 import android.os.Handler;
 import android.os.Looper;
-import android.support.annotation.ColorInt;
-import android.support.annotation.NonNull;
+
+import androidx.annotation.ColorInt;
+import androidx.annotation.NonNull;
 
 public class ProgressDrawable extends Drawable {
     private static final int MARQUEE_INTERVAL = 150;
@@ -25,7 +26,7 @@ public class ProgressDrawable extends Drawable {
 
     public ProgressDrawable() {
         mMarqueeLine = null;// marquee disabled
-        mMarqueeProgress = 0f;
+        mMarqueeProgress = 0.0f;
         mMarqueeSize = 0;
         mMarqueeNext = 0;
     }
@@ -35,8 +36,8 @@ public class ProgressDrawable extends Drawable {
         mForegroundColor = foreground;
     }
 
-    public void setProgress(float progress) {
-        mProgress = progress;
+    public void setProgress(double progress) {
+        mProgress = (float) progress;
         invalidateSelf();
     }
 
@@ -121,7 +122,7 @@ public class ProgressDrawable extends Drawable {
     }
 
     private void setupMarquee(int width, int height) {
-        mMarqueeSize = (int) ((width * 10f) / 100f);// the size is 10% of the width
+        mMarqueeSize = (int) ((width * 10.0f) / 100.0f);// the size is 10% of the width
 
         mMarqueeLine.rewind();
         mMarqueeLine.moveTo(-mMarqueeSize, -mMarqueeSize);
